requirements.txt
attrs==21.4.0
click==8.1.2
Faker==13.4.0
Flask==2.1.1
flask-paginate==2022.1.8
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
